Fix issue where calling Exit() hangs or throws InvalidOperationException
Just calling exit causes the monogame app to hang (unknown reason).
SDL.SDL_Quit() needs to be called to actually quit. However calling this
causes a System.InvalidOperationException to be thrown as the system is
still trying to draw to a destroyed window. This is fixed by calling
Game.SupressDraw() before we try and quit.